% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterative_predict.R
\name{i_predict}
\alias{i_predict}
\title{Iteratively predicted values based on a model object}
\usage{
i_predict(object, newdata, ...)
}
\arguments{
\item{object}{a model object for predictions.}

\item{...}{additional arguments affecting the predictions.}
}
\value{
The form of the value returned by \code{i_predict} depends on the class of its argument. See details of the particular methods for details of what is produced by that method.
}
\description{
Function to iteratively predict/forecast from specified \code{model} object, primarily for predicting models that include lagged dependent variables
}
\details{
This function provides a general approach to the problem of producing forecasts, or out of sample predictions, from (linear) models containing lagged dependent variables.  Other approaches to producing rolling or iterative forecasts include: \href{https://www.r-bloggers.com/variations-on-rolling-forecasts/}{Variations on rolling forecasts} and \href{https://stackoverflow.com/questions/4856555/iteratively-forecasting-dyn-models}{Iteratively forecasting dyn models}.

Most prediction methods that are similar to those for linear models have an argument \code{newdata} specifying the first place to look for explanatory variables to be used for prediction. Some considerable attempts are made to match up the columns in \code{newdata} to those used for fitting, for example that they are of comparable types and that any factors have the same level set in the same order (or can be transformed to be so).
}
\note{
Function not yet tested against all possible \code{predict} methods.
}
\examples{
 data(AvData);
 lm.AirFrgt <- lm(log(Freight.TKM) ~ log(A2304402X) + log(Aircraft.Departures)
                  + dQ2 + dQ3 + dQ4 + dPilot.Sep89 + dPilot.Dec89 + dPilot.Mar90
                  + lag(log(Freight.TKM)), data = AvData);
 data(FcstData);
 FcstData <- transform(FcstData,
                       A2304113C = splice_series(A2304113C, Fcst_Real_GDP_level),
                       A2304402X = splice_series(A2304402X, Fcst_Real_GDP_level),
                       Aircraft.Departures = splice_series(Aircraft.Departures, Fcst_Aircraft.Departures));
 AvFcsts <- exp(i_predict(lm.AirFrgt, newdata=Fcst.AvData, interval="prediction"));
 tail(AvFcsts);
 
}
\seealso{
\code{predict}, \code{predict.lm}, \code{predict.glm}.
}
\author{
David Mitchell <david.p.mitchell@homemail.com.au>
}
