% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterative_predict.R
\name{inverse_fn}
\alias{inverse_fn}
\title{Return specified function inverse}
\usage{
inverse_fn(x)
}
\arguments{
\item{x}{valid built-in function character string or mathematical operation.}
}
\value{
Returns the inverse of the function specified in \code{x} as a character string
}
\description{
Function to return the inverse of the supplied function
}
\note{
Function explicitly defined for use in \code{i_predict}. Function tested against R built-in functions (e.g. log(), sin(), etc.), but not yet thoroughly tested for mathematical expressions (e.g. x'^2' or '2*'x).
}
\examples{
 fn <- "log()"
 inverse_fn(fn)
}
\seealso{
{i_predict}
}
\author{
David Mitchell <david.p.mitchell@homemail.com.au>
}
