% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.utils.R
\name{splice_series}
\alias{splice_series}
\title{Splice two of more vectors into a single series}
\usage{
splice_series(x, y, base.period = NULL, method = "vector",
  direction = "forward")
}
\arguments{
\item{x}{Original (base) vector}

\item{y}{New or future vector series to be applied to the base vector (\code{x})}

\item{base.period}{Base period (vector index) on which to splice the two series.  If NULL, then function replaces all NA values in \code{x} using values in \code{y}. (Currently unimplemented.)}

\item{method}{One of either 'vector' (default) or 'iterative'.}

\item{direction}{One of either 'forward' (default) or 'backward'. (Currently not implemented.)}
}
\value{
An object of the same class as object \code{x}.
}
\description{
Function to iteratively splice two vectors, covering different bases, into series  extend
}
\note{
Arguments \code{base.period} and \code{direction} are currently unimplemented.
}
\examples{
  A <- ts(c(100, 95, 125, 150,  NA,  NA,  NA,  NA), start=1991);
  B <- ts(c(NA,  NA,  NA, 100, 120, 150, 200, 225), start=1991);
  splice_series(A, B);
}
\author{
Dave Mitchell <david.p.mitchell@homemail.com.au>
}
